requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
anywidget==0.10.0
asgiref==3.11.1
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
chatlas==0.15.2
chevron==0.14.0
click==8.3.2
commonmark==0.9.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
faicons==0.2.2
faiss-cpu==1.13.2
fastapi==0.135.3
Flask==3.1.3
frozenlist==1.8.0
fsspec==2024.9.0
great-tables==0.21.0
greenlet==3.4.0
h11==0.16.0
hf-xet==1.4.3
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.1
iniconfig==2.3.0
itsdangerous==2.2.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.28
langchain-huggingface==1.2.1
langsmith==0.7.30
linkify-it-py==2.1.0
matplotlib==3.10.8
mdit-py-plugins==0.5.0
multidict==6.7.1
multiprocess==0.70.16
munkres==1.1.4
nltk==3.9.4
openai==2.24.0
orjson==3.11.8
pandas==3.0.1
plotly==6.5.2
pluggy==1.6.0
propcache==0.4.1
psygnal==0.15.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
python-multipart==0.0.24
python-slugify==8.0.4
querychat==0.5.1
questionary==2.1.1
rank-bm25==0.2.2
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.8.0
sentence-transformers==5.4.0
setuptools==81.0.0
shellingham==1.5.4
shiny==1.5.1
shinychat==0.2.9
shinywidgets==0.7.1
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.5.3
typer==0.24.1
typing-inspection==0.4.2
uc-micro-py==2.0.0
uuid_utils==0.14.1
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.8
yarl==1.23.0
zstandard==0.25.0
appdirs==1.4.4
